reorganize items a little bit. also happymeal now uses the same script as toywrapper
