+++ zcrab/mcdonalds/items/food/appleslices/appleslices.consumable
