+++ zcrab/mcdonalds/items/food/appleslices/appleslices.png
