+++ zcrab/mcdonalds/items/food/chickennuggets/chickennuggets.consumable
