+++ zcrab/mcdonalds/items/food/chickennuggets/chickennuggets.png
