+++ zcrab/mcdonalds/items/food/chocolatemilk/chocolatemilk.consumable
