+++ zcrab/mcdonalds/items/food/chocolatemilk/chocolatemilk.png
