+++ zcrab/mcdonalds/items/food/drink/drink.png
