+++ zcrab/mcdonalds/items/food/emptywrapper/emptywrapper.consumable
