+++ zcrab/mcdonalds/items/food/emptywrapper/emptywrapper.png
