+++ zcrab/mcdonalds/items/food/fries/fries.consumable
