+++ zcrab/mcdonalds/items/food/fries/fries.png
