+++ zcrab/mcdonalds/items/happymeal/happymeal.animation
