+++ zcrab/mcdonalds/items/happymeal/mcdonald.ogg
