+++ zcrab/mcdonalds/items/treasurebag.lua
@@ -7,19 +7,21 @@ end
 function update(dt, fireMode, shiftHeld)
   updateAim()
 
-  if not storage.firing then return end
+  if not storage.firing then
+    return
+  end
 
   self.recoilRate = math.max(self.recoilRate + (10 * dt), 1)
   self.recoil = math.max(self.recoil - dt * self.recoilRate, 0)
 
   if storage.firing and animator.animationState("firing") == "off" then
     storage.firing = false
-    item.consume(1)
 
     if player then
       giveTreasure()
-      giveRandomToy()
     end
+    
+    item.consume(1)
   end
 end
 
@@ -48,19 +50,6 @@ function giveTreasure()
   local treasure = config.getParameter("treasure")
   local items = root.createTreasure(treasure.pool, treasure.level, treasure.seed)
   for _, item in pairs(items) do
-    player.giveItem(item)
-  end
-end
-
-function giveRandomToy()
-  local list = {}
-  local actionFigures = root.assetJson("/collections/actionfigures.collection:collectables")
-  for _, v in pairs(actionFigures) do
-    list[#list + 1] = v.item
-  end
-
-  local item = list[math.random(1, #list)]
-  if item then
     player.giveItem(item)
   end
 end
