+++ zcrab/mcdonalds/items/wrapper/toywrapper.animation
